added Time.cpp/.h
Implemented basic game loop

+++ src/GameWindow.h
@@ -8,6 +8,7 @@
 #include <memory>
 #include <SDL2/SDL.h>
 #include <iostream>
+#include <mach/i386/boolean.h>
 #include "Game.h"
 
 namespace GameEngine
@@ -16,16 +17,16 @@ namespace GameEngine
     {
     public:
         virtual ~GameWindow();
-        int getHeight();
-        int getWidth();
-        int getX();
-        int getY();
+        uint16_t getHeight();
+        uint16_t getWidth();
+        uint16_t getX();
+        uint16_t getY();
         std::string getTitle();
         virtual void start() = 0;
         virtual void stop() = 0;
         virtual void render(Game::Game *game) = 0;
     protected:
-        int width, height, x, y;
+        uint16_t width, height, x, y;
         std::string title;
     };
 
@@ -38,7 +39,7 @@ namespace GameEngine
     class SDLGameWindow : public GameWindow
     {
     public:
-        SDLGameWindow(std::string title, int width, int height);
+        SDLGameWindow(std::string title, uint16_t width, uint16_t height);
         ~SDLGameWindow();
         void render(Game::Game *game);
         void start();
